Remove doc about adding custom transforms
This doesn't work as well as it should given once you've piped
you're now back to the base `Stream` class.

The decorator version which adds methods at runtime is nicer for this
but still not type checkable.

For now simply using `pipe()` is probably fine until I can explore
this further. The direct parameter passing makes it nicer than
it was originally.